Create Spark_DF_Schema_Builder.py
# Description:
This script is used to handle illegal schema types not support in pyspark or Databricks environments. It will only convert illegal types to string but will keep every other accepted format in its native state. 

# Intended use:
This is mostly for working with parquet files but can also be used with CSV files.

# Requirements:
This script requires mount_point_builder to be run first to use specifically formatted mount point names to read from.